# docs/swagger.yaml
definitions:
...
  clientError.ErrorResponse:
    properties:
      error:
        type: string
      error_code:
        type: string
    type: object
...
  model.Moniest:
    properties:
      bio:
        type: string
      description:
        type: string
      id:
        type: string
      score:
        type: number
      subscription_info:
        $ref: '#/definitions/model.SubscriptionInfo'
    type: object
  model.OwnUser:
    properties:
      background_photo_link:
        type: string
      background_photo_thumbnail_link:
        type: string
      created_at:
        type: string
      email:
        type: string
      email_verified:
        type: boolean
      id:
        type: string
      location:
        type: string
      moniest:
        $ref: '#/definitions/model.Moniest'
      name:
        type: string
      profile_photo_link:
        type: string
      profile_photo_thumbnail_link:
        type: string
      surname:
        type: string
      updated_at:
        type: string
      username:
        type: string
    type: object
...
  model.SubscriptionInfo:
    properties:
      fee:
        type: number
      message:
        type: string
      updated_at:
        type: string
    type: object
...
  model.UpdateUserProfileRequest:
    properties:
      background_photo:
        type: string
      location:
        type: string
      name:
        type: string
      profile_photo:
        type: string
      surname:
        type: string
    type: object
  model.User:
    properties:
      background_photo_link:
        type: string
      background_photo_thumbnail_link:
        type: string
      created_at:
        type: string
      email_verified:
        type: boolean
      id:
        type: string
      location:
        type: string
      moniest:
        $ref: '#/definitions/model.Moniest'
      name:
        type: string
      profile_photo_link:
        type: string
      profile_photo_thumbnail_link:
        type: string
      surname:
        type: string
      updated_at:
        type: string
      username:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /users/:username:
    get:
      consumes:
      - application/json
      description: get user info with username
      parameters:
      - description: username
        in: path
        name: username
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '''email'' field will be visible if user request for own account'
          schema:
            $ref: '#/definitions/model.User'
        "404":
          description: not any user with this username
          schema:
            $ref: '#/definitions/clientError.ErrorResponse'
        "406":
          description: invalid username
          schema:
            $ref: '#/definitions/clientError.ErrorResponse'
        "500":
          description: server error
          schema:
            $ref: '#/definitions/clientError.ErrorResponse'
      security:
      - bearerAuth: []
      summary: Get User by Username
      tags:
      - User
  /users/:username/subscriptions:
    get:
      consumes:
      - application/json
      description: Get subscribed Users
      parameters:
      - description: username
        in: path
        name: username
        required: true
        type: string
      - description: 'default: 10 & max: 50'
        in: query
        name: limit
        required: true
        type: integer
      - description: 'default: 0'
        in: query
        name: offset
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.User'
            type: array
        "404":
          description: user not found
          schema:
            $ref: '#/definitions/clientError.ErrorResponse'
        "406":
          description: invalid body
          schema:
            $ref: '#/definitions/clientError.ErrorResponse'
        "500":
          description: server error
          schema:
            $ref: '#/definitions/clientError.ErrorResponse'
      security:
      - bearerAuth: []
      summary: Get Subscriptions
      tags:
      - User
  /users/profile:
    patch:
      consumes:
      - application/json
      description: Update user profile [name, surname, location, profile photo, background
        photo]
      parameters:
      - description: all fields are optional
        in: body
        name: UpdateUserBody
        required: true
        schema:
          $ref: '#/definitions/model.UpdateUserProfileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.OwnUser'
        "404":
          description: user not found
          schema:
            $ref: '#/definitions/clientError.ErrorResponse'
        "406":
          description: invalid body & data
          schema:
            $ref: '#/definitions/clientError.ErrorResponse'
        "500":
          description: server error
          schema:
            $ref: '#/definitions/clientError.ErrorResponse'
      security:
      - bearerAuth: []
      summary: Update User Profile
      tags:
      - User
securityDefinitions:
  bearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"